+++ tests/integration/test_oauth.py
@@ -4,9 +4,7 @@
 
 import logging
 import pathlib
-import subprocess
-from asyncio import gather, sleep
-from typing import Any, AsyncGenerator
+from asyncio import gather
 
 # import oauth_tools
 import pytest
@@ -21,13 +19,11 @@ from oauth_tools import (
 )
 from playwright.async_api._generated import Page
 from pytest_operator.plugin import OpsTest
-from tenacity import Retrying, stop_after_delay, wait_fixed
 
 pytest_plugins = ["oauth_tools.fixtures"]
 
 logger = logging.getLogger(__name__)
 
-MICROK8S_CLOUD_NAME = "uk8s"
 METADATA = yaml.safe_load(pathlib.Path("./metadata.yaml").read_text())
 APP_NAME = METADATA["name"]
 OPENSEARCH_APP_NAME = "opensearch"
@@ -48,102 +44,10 @@ DATA_INTEGRATOR_CONFIG = {
 }
 
 
-@pytest.fixture(scope="module")
-async def microk8s_cloud(ops_test: OpsTest) -> AsyncGenerator[None, Any]:
-    controller_name = next(
-        iter(yaml.safe_load(subprocess.check_output(["juju", "show-controller"])))
-    )
-
-    clouds = await ops_test._controller.clouds()
-    if f"cloud-{MICROK8S_CLOUD_NAME}" in clouds.clouds:
-        yield None
-        return
-
-    try:
-        subprocess.run(["sudo", "snap", "install", "--classic", "microk8s"], check=True)
-        subprocess.run(["sudo", "snap", "install", "--classic", "kubectl"], check=True)
-        subprocess.run(["sudo", "microk8s", "enable", "dns"], check=True)
-        subprocess.run(["sudo", "microk8s", "enable", "hostpath-storage"], check=True)
-        subprocess.run(
-            ["sudo", "microk8s", "enable", "metallb:10.64.140.43-10.64.140.49"],
-            check=True,
-        )
-
-        # Configure kubectl now
-        subprocess.run(["mkdir", "-p", str(pathlib.Path.home() / ".kube")], check=True)
-        kubeconfig = subprocess.check_output(["sudo", "microk8s", "config"])
-        with open(str(pathlib.Path.home() / ".kube" / "config"), "w") as f:
-            f.write(kubeconfig.decode())
-        for attempt in Retrying(stop=stop_after_delay(150), wait=wait_fixed(15)):
-            with attempt:
-                if (
-                    len(
-                        subprocess.check_output(
-                            "kubectl get po -A  --field-selector=status.phase!=Running",
-                            shell=True,
-                            stderr=subprocess.DEVNULL,
-                        ).decode()
-                    )
-                    != 0
-                ):  # We got sth different from "No resources found." in stderr
-                    raise Exception()
-
-        # Add microk8s to the kubeconfig
-        subprocess.run(
-            [
-                "juju",
-                "add-k8s",
-                MICROK8S_CLOUD_NAME,
-                "--client",
-                "--controller",
-                controller_name,
-            ],
-            check=True,
-        )
-    except subprocess.CalledProcessError as e:
-        pytest.exit(str(e))
-
-    yield None
-
-    if not ops_test.keep_model:
-        subprocess.run(
-            [
-                "juju",
-                "remove-cloud",
-                "--client",
-                "--controller",
-                controller_name,
-                MICROK8S_CLOUD_NAME,
-            ],
-            check=True,
-        )
-        subprocess.run(["sudo", "snap", "remove", "--purge", "microk8s"], check=True)
-        subprocess.run(["sudo", "snap", "remove", "--purge", "kubectl"], check=True)
-
-
-@pytest.fixture(scope="module")
-async def ops_test_microk8s(
-    request, tmp_path_factory, ops_test: OpsTest, microk8s_cloud: None
-) -> AsyncGenerator[OpsTest, Any]:
-    model_name = f"{ops_test.model_name}-uk8s"
-    request.config.option.controller = ops_test.controller_name
-    request.config.option.cloud = "uk8s"
-    request.config.option.model = model_name
-    request.config.option.model_alias = model_name
-    ops_res = OpsTest(request, tmp_path_factory)
-    await ops_res._setup_model()
-    yield ops_res
-    if not ops_test.keep_model:
-        await ops_res.forget_model(alias=model_name)
-        await ops_res._controller.destroy_model(model_name, destroy_storage=True, force=True)
-        while model_name in await ops_res._controller.list_models():
-            await sleep(5)
-    await ops_res._cleanup_models()
-
-
 @pytest.mark.abort_on_fail
 @pytest.mark.skip_if_deployed
 async def test_deploy(ops_test: OpsTest, ops_test_microk8s: OpsTest, charm: str, series: str):
+    """Deploy OpenSearch and OpenSearch Dashboards but don't wait for completion."""
     await ops_test.model.set_config(OPENSEARCH_CONFIG)
 
     await ops_test.model.deploy(
@@ -166,6 +70,7 @@ async def test_deploy(ops_test: OpsTest, ops_test_microk8s: OpsTest, charm: str,
 async def test_deploy_identity_bundle(
     ops_test: OpsTest, ops_test_microk8s: OpsTest, ext_idp_service: ExternalIdpService
 ):
+    """Deploy identity platform on K8s and wait for both models to complete deployments."""
     await deploy_identity_bundle(
         ops_test=ops_test_microk8s,
         bundle_channel="latest/edge",
@@ -180,6 +85,10 @@ async def test_deploy_identity_bundle(
 @pytest.mark.abort_on_fail
 @pytest.mark.skip_if_deployed
 async def test_setup_relations(ops_test: OpsTest, ops_test_microk8s: OpsTest):
+    """Establish all the required relations.
+
+    Connects OpenSearch, OpenSearch Dashboards and identity platform (cross-model).
+    """
     await ops_test_microk8s.model.create_offer(
         "certificates", "certificates", "self-signed-certificates"
     )
@@ -214,6 +123,7 @@ async def test_oauth(
     page: Page,
     ext_idp_service: ExternalIdpService,
 ):
+    """Ensure that SSO works for OpenSearch Dashboards login."""
     opensearch_dashboards_ip = await get_address(
         ops_test, ops_test.model.applications[APP_NAME].units[0].name
     )
